requirements.txt
google-adk
pandas
db-dtypes